PlayerHUD skills upgrade
To upgrade your skill click on the text of skill you want to upgrade (it is a beginning so u can do it a lot of times)

+++ 2DPlatformerController/Assets/Attributes/DamagerAttributes.cs
@@ -8,7 +8,7 @@ using System.Collections.Generic;
 [Serializable]
 public class DamagerAttributes
 {
-    public int AttackDamage;
+    public float AttackDamage;
     public float AttackCooldownInSeconds;
     public DateTime LastAttackTime;
     public float Radius;
